requirements.txt
Django==5.2.6
djangorestframework==3.16.1
python-dateutil==2.9.0.post0
requests==2.32.5
py3xui==0.5.1
jinja2==3.1.6
psycopg[binary]==3.2.3
pyTelegramBotAPI==4.29.1